BOAC-5884, initial pieces of PeerAdvisorHome and EditPeerAdvisingNote
