+++ src/components/peer/note/EditPeerAdvisingNoteHeader.vue
@@ -1,0 +1,16 @@
+<template>
+  <div class="d-flex justify-space-between">
+    <div>
+      <ModalHeader text="New Note" />
+    </div>
+    <div>
+      <SelectNoteTemplateForNote :exit="noop" />
+    </div>
+  </div>
+</template>
+
+<script setup lang="ts">
+import {noop} from 'lodash'
+import ModalHeader from '@/components/util/ModalHeader.vue'
+import SelectNoteTemplateForNote from '@/components/note/template/SelectNoteTemplateForNote.vue'
+</script>
